dom/base/test/test_anonymousContent_xul_window.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1020244
Check that XUL windows aren't supported, that the API throws, but doesn't crash.
-->
<window title="Anonymous content in a XUL window"
xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml"></body>
  <box>This is a test box</box>

  <script type="application/javascript;version=1.8">
    // Insert content
    let testElement = document.querySelector("box");

    let didThrow = false;
    try {
      document.insertAnonymousContent(testElement);
    } catch (e) {
      didThrow = true;
    }

    ok(didThrow,
      "Inserting anonymous content in a XUL document did throw an exception")
  </script>
</window>
